Second version of the Second Life MitM proxy.
According to my notes, in this version the packet parser was finally complete.

Back then the Second Life protocol was UDP based. Parsing wasn't hardcoded into the scripts, but rather the packet format was encoded in a text file that came with the Second Life client source code. That means upgrading to a new version mostly just meant replacing that text file.